[win10] don't try to resize dx device at exit

+++ xbmc/windowing/win10/WinSystemWin10.h
@@ -72,7 +72,7 @@ struct MONITOR_DETAILS
   // Windows desktop info
   int       ScreenWidth;
   int       ScreenHeight;
-  int       RefreshRate;
+  float     RefreshRate;
   int       Bpp;
   bool      Interlaced;
